requirements/base.txt
Django==3.2.25
django-split-settings==1.2.0
django-environ==0.11.2
django-mptt==0.16.0
djangorestframework==3.15.1
